requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2

# For Azure identity integration (optional)
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0